Refactor interaction classes into single file
